requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-telegram-bot==21.0.1
python-dotenv==1.0.1